added TODOs to client code

+++ Server_arm/Middlewares/ST/STM32_WPAN/ble/svc/Inc/p2p_stm.h
@@ -42,7 +42,7 @@ typedef enum
 
 typedef struct
 {
-  uint8_t * pPayload;
+  uint16_t * pPayload;
   uint8_t     Length;
 }P2PS_STM_Data_t;
